finish converting test_suffix to regex and adding tests for different runners
